+++ tests/test_fsoe_master.py
@@ -3,7 +3,13 @@ from typing import TYPE_CHECKING
 import pytest
 
 from ingeniamotion.enums import FSoEState
-from ingeniamotion.fsoe import FSOE_MASTER_INSTALLED, FSoEError, FSoEMaster
+from ingeniamotion.fsoe import (
+    FSOE_MASTER_INSTALLED,
+    FSoEError,
+    SafeInputsFunction,
+    SS1Function,
+    STOFunction,
+)
 from ingeniamotion.motion_controller import MotionController
 from tests.conftest import timeout_loop
 
@@ -48,14 +54,6 @@ def error_handler(error: FSoEError):
     raise RuntimeError(f"FSoE error received: {error}")
 
 
-@pytest.mark.fsoe
-def test_fsoe_master_get_application_parameters(mc, alias):
-    assert isinstance(mc.fsoe, FSoEMaster)
-
-    application_parameters = mc.fsoe._get_application_parameters(servo=alias)
-    assert len(application_parameters)
-
-
 @pytest.fixture()
 def mc_with_fsoe(mc):
     # Subscribe to emergency messages
@@ -63,16 +61,75 @@ def mc_with_fsoe(mc):
     # Configure error channel
     mc.fsoe.subscribe_to_errors(error_handler)
     # Create and start the FSoE master handler
-    mc.fsoe.create_fsoe_master_handler()
-    yield mc
+    handler = mc.fsoe.create_fsoe_master_handler()
+    yield mc, handler
     # IM should be notified and clear references when a servo is disconnected from ingenialink
     # https://novantamotion.atlassian.net/browse/INGM-624
     mc.fsoe._delete_master_handler()
 
 
+@pytest.mark.fsoe
+def test_fsoe_master_get_safety_parameters(mc_with_fsoe):
+    mc, handler = mc_with_fsoe
+
+    assert len(handler.safety_parameters) != 0
+
+
+@pytest.mark.fsoe
+def test_mandatory_safety_functions(mc_with_fsoe):
+    mc, handler = mc_with_fsoe
+
+    safety_functions_by_types = handler.safety_functions_by_type()
+
+    sto_instances = safety_functions_by_types[STOFunction]
+    assert len(sto_instances) == 1
+
+    ss1_instances = safety_functions_by_types[SS1Function]
+    assert len(ss1_instances) >= 1
+
+    si_instances = safety_functions_by_types[SafeInputsFunction]
+    assert len(si_instances) == 1
+
+
+@pytest.mark.fsoe
+def test_getter_of_safety_functions(mc_with_fsoe):
+    mc, handler = mc_with_fsoe
+
+    sto_function = STOFunction(command=None, io=None, parameters=None)
+    ss1_function_1 = SS1Function(command=None, time_to_sto=None, io=None, parameters=None)
+    ss1_function_2 = SS1Function(command=None, time_to_sto=None, io=None, parameters=None)
+
+    handler.safety_functions = (sto_function, ss1_function_1, ss1_function_2)
+    handler.get_function_instance.cache_clear()
+
+    # Single instance of STOFunction
+    assert handler.get_function_instance(STOFunction) is sto_function
+    assert handler.get_function_instance(STOFunction, instance=1) is sto_function
+
+    # Multiple instances of SS1Function
+    with pytest.raises(ValueError) as error:
+        # Must specify the instance
+        handler.get_function_instance(SS1Function)
+    assert (
+        error.value.args[0]
+        == "Multiple SS1Function instances found (2). Specify the instance number."
+    )
+
+    with pytest.raises(IndexError) as error:
+        # Instance 0 does not exist
+        handler.get_function_instance(SS1Function, instance=0)
+    assert error.value.args[0] == "Master handler does not contain SS1Function instance 0"
+    assert handler.get_function_instance(SS1Function, instance=1) is ss1_function_1
+    assert handler.get_function_instance(SS1Function, instance=2) is ss1_function_2
+    with pytest.raises(IndexError) as error:
+        # Instance 3 does not exist
+        handler.get_function_instance(SS1Function, instance=3)
+    assert error.value.args[0] == "Master handler does not contain SS1Function instance 3"
+
+
 @pytest.fixture()
 def mc_state_data(mc_with_fsoe):
-    mc = mc_with_fsoe
+    mc, handler = mc_with_fsoe
 
     mc.fsoe.configure_pdos(start_pdos=True)
     # Wait for the master to reach the Data state
@@ -96,7 +153,7 @@ def test_safe_inputs_value(mc_state_data):
 
 @pytest.mark.fsoe
 def test_safety_address(mc_with_fsoe, alias):
-    mc = mc_with_fsoe
+    mc, handler = mc_with_fsoe
 
     master_handler = mc.fsoe._handlers[alias]
 
@@ -133,7 +190,7 @@ def mc_state_to_fsoe_master_state(state: FSoEState):
     ],
 )
 def test_get_master_state(mocker, mc_with_fsoe, state_enum):
-    mc = mc_with_fsoe
+    mc, handler = mc_with_fsoe
 
     # Master state is obtained as function
     # and not on the parametrize