+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# Unreleased
+### Added
+- Added Model for safety functions and safe parameters.
+
 ## [0.9.1] - 2025-05-07
 ### Added
 - Methods to scan Ethernet drives.